Put transfer of cache file in lazy-loading
Here, we allow the transfer of a cache db file to be initiated when an
object in that cache file is first accessed (rather than transfer the
cache file *before* the object is accessed, which is what we did before).
This requires 'cacheLazyLoad' to be knowledgeable of the internals of
the cache directory, but oh well....

Also, the cache file format has changed so that the index and offset
can be read remotely.