Fixes to 'checkcode()'
First, we need to load the 'checkenv' environment *per expression*; we
can't just load everything at the beginning because objects with the
same names in different expressions will overwrite each other.

Second, when an evaluation fails and we try to lazy-load from the cache,
we need to force a download of an object (if the cache has been
cloned).  For some reason if this isn't done, then the object fails
to be verified.