Change cacheLazyLoad logic
When trying to find a db file in 'cacheLazyLoad'; if it exists,
then just lazy-load it.  If it doesn't exist and we are a clone,
then try to transfer it from the remote.  If it doesn't exist and
we're not a clone, then just throw an error.

The previous logic did not distinguish between being a clone
and not being a clone.