Copy objects to globalenv on first eval
We used to evaluate an expression for the first time, cache it
to the database, and the lazy-load it back into the global
environment.  This is obviously wasteful because if we need the objects
from the evaluation, we have to load them from the disk.  So on first
evaluation, just copy the objects from the temporary environment
to the global environment.