Remove dependencies on 'digest' and 'tools'
We now use 'sha1_file' and 'sha1' to hash files and objects.  The
behavior should currently be identical to what we used to have when
we were using the 'digest' package.  However, in the 'identicalFiles'
function, we now use the 'sha1_file' function instead of the
'md5sum' function from the 'tools' package to compare file contents.